+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,39 +1,16 @@
 package roomescape.repository;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicLong;
-import org.springframework.stereotype.Repository;
 import roomescape.entity.Reservation;
 
-@Repository
-public class ReservationRepository {
+public interface ReservationRepository {
 
-    private final List<Reservation> reservations = Collections.synchronizedList(new ArrayList<>());
-    private final AtomicLong index = new AtomicLong(1L);
+    List<Reservation> findAll();
 
-    public long getNextId() {
-        return index.getAndIncrement();
-    }
+    Optional<Reservation> findById(Long id);
 
-    public List<Reservation> findAll() {
-        return new ArrayList<>(reservations);
-    }
+    Reservation save(Reservation reservation);
 
-    public Optional<Reservation> findById(Long id) {
-        return reservations.stream()
-                .filter(reservation -> reservation.getId().equals(id))
-                .findFirst();
-    }
-
-    public Reservation save(Reservation reservation) {
-        reservations.add(reservation);
-        return reservation;
-    }
-
-    public void delete(Reservation reservation) {
-        reservations.remove(reservation);
-    }
+    void delete(Reservation reservation);
 }
