+++ src/main/java/roomescape/repository/ReservationRepositoryImpl.java
@@ -1,0 +1,42 @@
+package roomescape.repository;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicLong;
+import org.springframework.stereotype.Repository;
+import roomescape.entity.Reservation;
+
+@Repository
+public class ReservationRepositoryImpl implements ReservationRepository {
+
+    private final List<Reservation> reservations = Collections.synchronizedList(new ArrayList<>());
+    private final AtomicLong index = new AtomicLong(1L);
+
+    @Override
+    public List<Reservation> findAll() {
+        return new ArrayList<>(reservations);
+    }
+
+    @Override
+    public Optional<Reservation> findById(Long id) {
+        return reservations.stream()
+                .filter(reservation -> reservation.getId().equals(id))
+                .findFirst();
+    }
+
+    @Override
+    public Reservation save(Reservation reservation) {
+        if (reservation.getId() == null) {
+            reservation.setId(index.getAndIncrement());
+        }
+        reservations.add(reservation);
+        return reservation;
+    }
+
+    @Override
+    public void delete(Reservation reservation) {
+        reservations.remove(reservation);
+    }
+}
